+++ src/lib/libast/include/swap.h
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2022 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 /*
@@ -36,8 +37,8 @@
 #define SWAPOP(n)	(((n)&int_swap)^(n))
 
 extern void*		swapmem(int, const void*, void*, size_t);
-extern intmax_t		swapget(int, const void*, int);
-extern void*		swapput(int, void*, int, intmax_t);
-extern int		swapop(const void*, const void*, int);
+extern intmax_t		swapget(int, const void*, size_t);
+extern void*		swapput(int, void*, size_t, intmax_t);
+extern ssize_t		swapop(const void*, const void*, ssize_t);
 
 #endif
